skills/image-upscaler/pyproject.toml
[project]
name = "skill-image-upscaler"
version = "0.1.0"
description = "Losslessly upscale images to a target file size using Pillow and Lanczos resampling"
readme = "SKILL.md"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0.0",
]

[tool.uv]
package = false
